+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Note
 
+## v2.6.0
+
+- Added `FilteredStreamResponse` as the return type of `Filtered Stream` endpoint. And now we can see the IDs and tags of the rules matched in the streaming from the `FilteredStreamResponse.matchingRules` field. The `FilteredStreamResponse` extends `TwitterResponse`, so there are no destructive changes. ([#226](https://github.com/twitter-dart/twitter-api-v2/issues/226))
+
 ## v2.5.0
 
 - Added `spaceFields` argument. ([#184](https://github.com/twitter-dart/twitter-api-v2/issues/184))